Trait ndarray_linalg::cholesky::CholeskyInplace

pub trait CholeskyInplace {
    fn cholesky_inplace(&mut self, uplo: UPLO) -> Result<&mut Self>;
}

Cholesky decomposition of Hermitian (or real symmetric) positive definite mutable reference of matrix

Required methods

fn cholesky_inplace(&mut self, uplo: UPLO) -> Result<&mut Self>

Computes the Cholesky decomposition of the Hermitian (or real symmetric) positive definite matrix, writing the result (L or U according to the argument) to self and returning it.

If the argument is UPLO::Upper, then computes the decomposition A = U^H * U using the upper triangular portion of A and writes U. Otherwise, if the argument is UPLO::Lower, computes the decomposition A = L * L^H using the lower triangular portion of A and writes L.

Loading content...