Trait ndarray_linalg::cholesky::FactorizeC

pub trait FactorizeC<S: Data> {
    // Required method
    fn factorizec(&self, uplo: UPLO) -> Result<CholeskyFactorized<S>>;
}

Cholesky decomposition of Hermitian (or real symmetric) positive definite matrix reference

Required Methods

fn factorizec(&self, uplo: UPLO) -> Result<CholeskyFactorized<S>>

Computes the Cholesky decomposition of the Hermitian (or real symmetric) positive definite matrix.

If the argument is UPLO::Upper, then computes the decomposition A = U^H * U using the upper triangular portion of A and returns the factorization containing U. Otherwise, if the argument is UPLO::Lower, computes the decomposition A = L * L^H using the lower triangular portion of A and returns the factorization containing L.