ndarray_linalg
Module krylov

Krylov subspace methods

Re-exports

pub use arnoldi::arnoldi_householder;

pub use arnoldi::arnoldi_mgs;

pub use arnoldi::Arnoldi;

pub use householder::householder;

pub use householder::Householder;

pub use mgs::mgs;

pub use mgs::MGS;

Modules

arnoldi

Arnoldi iteration

householder

Householder reflection

mgs

Modified Gram-Schmit orthogonalizer

Enums

AppendResult

Strategy

Strategy for linearly dependent vectors appearing in iterative QR decomposition

Traits

Orthogonalizer

Trait for creating orthogonal basis from iterator of arrays

Functions

qr

Online QR decomposition using arbitrary orthogonalizer

Type Aliases

Coefficients

Array type for coefficients to the current basis

H

H-matrix

Q

Q-matrix

R

R-matrix