Trait ndarray_linalg::lapack::solve::Solve_

pub trait Solve_: Scalar + Sized {
    unsafe fn lu(l: MatrixLayout, a: &mut [Self]) -> Result<Pivot>;

    unsafe fn inv(l: MatrixLayout, a: &mut [Self], p: &Pivot) -> Result<()>;

    unsafe fn rcond(
        l: MatrixLayout, 
        a: &[Self], 
        anorm: Self::Real
    ) -> Result<Self::Real>;

    unsafe fn solve(
        l: MatrixLayout, 
        t: Transpose, 
        a: &[Self], 
        p: &Pivot, 
        b: &mut [Self]
    ) -> Result<()>;
}

Wraps *getrf, *getri, and *getrs

Required methods

unsafe fn lu(l: MatrixLayout, a: &mut [Self]) -> Result<Pivot>

Computes the LU factorization of a general m x n matrix a using partial pivoting with row interchanges.

If the result matches Err(LinalgError::Lapack(LapackError { return_code )) if return_code > 0, then U[(return_code-1, return_code-1)] is exactly zero. The factorization has been completed, but the factor U is exactly singular, and division by zero will occur if it is used to solve a system of equations.

unsafe fn inv(l: MatrixLayout, a: &mut [Self], p: &Pivot) -> Result<()>

unsafe fn rcond(
l: MatrixLayout,
a: &[Self],
anorm: Self::Real
) -> Result<Self::Real>

Estimates the the reciprocal of the condition number of the matrix in 1-norm.

anorm should be the 1-norm of the matrix a.

unsafe fn solve(
l: MatrixLayout,
t: Transpose,
a: &[Self],
p: &Pivot,
b: &mut [Self]
) -> Result<()>

Loading content...