Struct ndarray_linalg::solve::LUFactorized

pub struct LUFactorized<S: Data + RawDataClone> {
    pub a: ArrayBase<S, Ix2>,
    pub ipiv: Pivot,
}

Represents the LU factorization of a matrix A as A = P*L*U.

Fields
a: ArrayBase<S, Ix2>

The factors L and U; the unit diagonal elements of L are not stored.

ipiv: Pivot

The pivot indices that define the permutation matrix P.