ndarray_linalg::solveh

Struct BKFactorized

pub struct BKFactorized<S: Data> {
    pub a: ArrayBase<S, Ix2>,
    pub ipiv: Pivot,
}

Represents the Bunch–Kaufman factorization of a Hermitian (or real symmetric) matrix as A = P * U * D * U^H * P^T.

Fields
a: ArrayBase<S, Ix2>ipiv: Pivot
Implementations

impl<A, S> BKFactorized<S>
where A: Scalar + Lapack, S: Data<Elem = A>,

pub fn deth(&self) -> A::Real

Computes the determinant of the factorized Hermitian (or real symmetric) matrix.

pub fn sln_deth(&self) -> (A::Real, A::Real)

Computes the (sign, natural_log) of the determinant of the factorized Hermitian (or real symmetric) matrix.

The natural_log is the natural logarithm of the absolute value of the determinant. If the determinant is zero, sign is 0 and natural_log is negative infinity.

To obtain the determinant, you can compute sign * natural_log.exp() or just call .deth() instead.

This method is more robust than .deth() to very small or very large determinants since it returns the natural logarithm of the determinant rather than the determinant itself.

pub fn deth_into(self) -> A::Real

Computes the determinant of the factorized Hermitian (or real symmetric) matrix.

pub fn sln_deth_into(self) -> (A::Real, A::Real)

Computes the (sign, natural_log) of the determinant of the factorized Hermitian (or real symmetric) matrix.

The natural_log is the natural logarithm of the absolute value of the determinant. If the determinant is zero, sign is 0 and natural_log is negative infinity.

To obtain the determinant, you can compute sign * natural_log.exp() or just call .deth_into() instead.

This method is more robust than .deth_into() to very small or very large determinants since it returns the natural logarithm of the determinant rather than the determinant itself.