Struct ndarray_linalg::TruncatedEig

pub struct TruncatedEig<A: Scalar> {
    pub constraints: Option<Array2<A>>,
    // some fields omitted
}

Truncated eigenproblem solver

This struct wraps the LOBPCG algorithm and provides convenient builder-pattern access to parameter like maximal iteration, precision and constraint matrix. Furthermore it allows conversion into a iterative solver where each iteration step yields a new eigenvalue/vector pair.

Fields
constraints: Option<Array2<A>>
Implementations

impl<A: Float + Scalar + ScalarOperand + Lapack + PartialOrd + Default> TruncatedEig<A>

pub fn new(problem: Array2<A>, order: Order) -> TruncatedEig<A>

pub fn precision(self, precision: f32) -> Self

pub fn maxiter(self, maxiter: usize) -> Self

pub fn orthogonal_to(self, constraints: Array2<A>) -> Self

pub fn precondition_with(self, preconditioner: Array2<A>) -> Self

pub fn decompose(&self, num: usize) -> LobpcgResult<A>