Trait ndarray_linalg::tridiagonal::ReciprocalConditionNumTridiagonalInto

pub trait ReciprocalConditionNumTridiagonalInto<A: Scalar> {
    fn rcond_tridiagonal_into(self) -> Result<A::Real>;
}

An interface for estimating the reciprocal condition number of tridiagonal matrices.

Required methods

fn rcond_tridiagonal_into(self) -> Result<A::Real>

Estimates the reciprocal of the condition number of the tridiagonal matrix in 1-norm.

This method uses the LAPACK *gtcon routines, which estimate self.inv_tridiagonal().opnorm_one() and then compute rcond = 1. / (self.opnorm_one() * self.inv_tridiagonal().opnorm_one()).

If rcond is near 0., the matrix is badly conditioned.

If rcond is near 1., the matrix is well conditioned.

Loading content...