Trait ndarray_linalg::cholesky::DeterminantCInto

pub trait DeterminantCInto {
    type Output;
    fn detc_into(self) -> Self::Output;

    fn ln_detc_into(self) -> Self::Output;
}

Determinant of Hermitian (or real symmetric) positive definite matrix

Associated Types

type Output

Loading content...
Required methods

fn detc_into(self) -> Self::Output

Computes the determinant of the Hermitian (or real symmetric) positive definite matrix.

fn ln_detc_into(self) -> Self::Output

Computes the natural log of the determinant of the Hermitian (or real symmetric) positive definite matrix.

This method is more robust than .detc_into() to very small or very large determinants since it returns the natural logarithm of the determinant rather than the determinant itself.

Loading content...