Trait ndarray_linalg::cholesky::SolveC

pub trait SolveC<A: Scalar> {
    // Required method
    fn solvec_inplace<'a, S: DataMut<Elem = A>>(
        &self,
        b: &'a mut ArrayBase<S, Ix1>,
    ) -> Result<&'a mut ArrayBase<S, Ix1>>;

    // Provided methods
    fn solvec<S: Data<Elem = A>>(
        &self,
        b: &ArrayBase<S, Ix1>,
    ) -> Result<Array1<A>> { ... }
    fn solvec_into<S: DataMut<Elem = A>>(
        &self,
        b: ArrayBase<S, Ix1>,
    ) -> Result<ArrayBase<S, Ix1>> { ... }
}

Solve systems of linear equations with Hermitian (or real symmetric) positive definite coefficient matrices

Required Methods

fn solvec_inplace<'a, S: DataMut<Elem = A>>( &self, b: &'a mut ArrayBase<S, Ix1>, ) -> Result<&'a mut ArrayBase<S, Ix1>>

Solves a system of linear equations A * x = b with Hermitian (or real symmetric) positive definite matrix A, where A is self, b is the argument, and x is the successful result. The value of x is also assigned to the argument.

Provided Methods

fn solvec<S: Data<Elem = A>>(&self, b: &ArrayBase<S, Ix1>) -> Result<Array1<A>>

Solves a system of linear equations A * x = b with Hermitian (or real symmetric) positive definite matrix A, where A is self, b is the argument, and x is the successful result.

fn solvec_into<S: DataMut<Elem = A>>( &self, b: ArrayBase<S, Ix1>, ) -> Result<ArrayBase<S, Ix1>>

Solves a system of linear equations A * x = b with Hermitian (or real symmetric) positive definite matrix A, where A is self, b is the argument, and x is the successful result.

Object Safety

This trait is not object safe.