Trait ndarray_linalg::least_squares::LeastSquaresSvd

pub trait LeastSquaresSvd<D, E, I> where
    D: Data<Elem = E>,
    E: Scalar + Lapack,
    I: Dimension, {
    fn least_squares(
        &self, 
        rhs: &ArrayBase<D, I>
    ) -> Result<LeastSquaresResult<E, I>>;
}

Solve least squares for immutable references

Required methods

fn least_squares(
&self,
rhs: &ArrayBase<D, I>
) -> Result<LeastSquaresResult<E, I>>

Solve a least squares problem of the form Ax = rhs by calling A.least_squares(&rhs). A and rhs are unchanged.

A and rhs must have the same layout, i.e. they must be both either row- or column-major format, otherwise a IncompatibleShape error is raised.

Loading content...