Trait ndarray_linalg::least_squares::LeastSquaresSvdInPlace

pub trait LeastSquaresSvdInPlace<D, E, I> where
    D: Data<Elem = E>,
    E: Scalar + Lapack,
    I: Dimension, {
    fn least_squares_in_place(
        &mut self, 
        rhs: &mut ArrayBase<D, I>
    ) -> Result<LeastSquaresResult<E, I>>;
}

Solve least squares for mutable references, overwriting the input fields in the process

Required Methods

fn least_squares_in_place(
&mut self,
rhs: &mut ArrayBase<D, I>
) -> Result<LeastSquaresResult<E, I>>

Solve a least squares problem of the form Ax = rhs by calling A.least_squares(&mut rhs), overwriting both A and rhs. This uses the memory location of A and rhs, which avoids some extra memory allocations.

A and rhs must have the same layout, i.e. they must be both either row- or column-major format, otherwise a IncompatibleShape error is raised.