Module ndarray_linalg::tridiagonal

Vectors as a Tridiagonal matrix & Methods for tridiagonal matrices

Structs

LUFactorizedTridiagonal

Represents the LU factorization of a tridiagonal matrix A as A = P*L*U.

Tridiagonal

Represents a tridiagonal matrix as 3 one-dimensional vectors.

Traits

DeterminantTridiagonal

An interface for calculating determinants of tridiagonal matrix refs.

ExtractTridiagonal

An interface for making a Tridiagonal struct.

FactorizeTridiagonal

An interface for computing LU factorizations of tridiagonal matrix refs.

FactorizeTridiagonalInto

An interface for computing LU factorizations of tridiagonal matrices.

ReciprocalConditionNumTridiagonal

An interface for estimating the reciprocal condition number of tridiagonal matrix refs.

ReciprocalConditionNumTridiagonalInto

An interface for estimating the reciprocal condition number of tridiagonal matrices.

SolveTridiagonal

SolveTridiagonalInplace