Struct ndarray_linalg::tridiagonal::LUFactorizedTridiagonal

pub struct LUFactorizedTridiagonal<A> where
    A: Scalar,  {
    pub a: Tridiagonal<A>,
    pub du2: Vec<A>,
    pub ipiv: Vec<i32>,
    // some fields omitted
}

Represents the LU factorization of a tridiagonal matrix A as A = P*L*U.

Fields
a: Tridiagonal<A>

A tridiagonal matrix which consists of

l : layout of raw matrix

dl: (n-1) multipliers that define the matrix L.

d : (n) diagonal elements of the upper triangular matrix U.

du: (n-1) elements of the first super-diagonal of U.

du2: Vec<A>

(n-2) elements of the second super-diagonal of U.

ipiv: Vec<i32>

The pivot indices that define the permutation matrix P.