ndarray_linalg::tridiagonal

Trait SolveTridiagonal

pub trait SolveTridiagonal<A: Scalar, D: Dimension> {
    // Required methods
    fn solve_tridiagonal<S: Data<Elem = A>>(
        &self,
        b: &ArrayBase<S, D>,
    ) -> Result<Array<A, D>>;
    fn solve_tridiagonal_into<S: DataMut<Elem = A>>(
        &self,
        b: ArrayBase<S, D>,
    ) -> Result<ArrayBase<S, D>>;
    fn solve_t_tridiagonal<S: Data<Elem = A>>(
        &self,
        b: &ArrayBase<S, D>,
    ) -> Result<Array<A, D>>;
    fn solve_t_tridiagonal_into<S: DataMut<Elem = A>>(
        &self,
        b: ArrayBase<S, D>,
    ) -> Result<ArrayBase<S, D>>;
    fn solve_h_tridiagonal<S: Data<Elem = A>>(
        &self,
        b: &ArrayBase<S, D>,
    ) -> Result<Array<A, D>>;
    fn solve_h_tridiagonal_into<S: DataMut<Elem = A>>(
        &self,
        b: ArrayBase<S, D>,
    ) -> Result<ArrayBase<S, D>>;
}

Required Methods

fn solve_tridiagonal<S: Data<Elem = A>>( &self, b: &ArrayBase<S, D>, ) -> Result<Array<A, D>>

Solves a system of linear equations A * x = b with tridiagonal matrix A, where A is self, b is the argument, and x is the successful result.

fn solve_tridiagonal_into<S: DataMut<Elem = A>>( &self, b: ArrayBase<S, D>, ) -> Result<ArrayBase<S, D>>

Solves a system of linear equations A * x = b with tridiagonal matrix A, where A is self, b is the argument, and x is the successful result.

fn solve_t_tridiagonal<S: Data<Elem = A>>( &self, b: &ArrayBase<S, D>, ) -> Result<Array<A, D>>

Solves a system of linear equations A^T * x = b with tridiagonal matrix A, where A is self, b is the argument, and x is the successful result.

fn solve_t_tridiagonal_into<S: DataMut<Elem = A>>( &self, b: ArrayBase<S, D>, ) -> Result<ArrayBase<S, D>>

Solves a system of linear equations A^T * x = b with tridiagonal matrix A, where A is self, b is the argument, and x is the successful result.

fn solve_h_tridiagonal<S: Data<Elem = A>>( &self, b: &ArrayBase<S, D>, ) -> Result<Array<A, D>>

Solves a system of linear equations A^H * x = b with tridiagonal matrix A, where A is self, b is the argument, and x is the successful result.

fn solve_h_tridiagonal_into<S: DataMut<Elem = A>>( &self, b: ArrayBase<S, D>, ) -> Result<ArrayBase<S, D>>

Solves a system of linear equations A^H * x = b with tridiagonal matrix A, where A is self, b is the argument, and x is the successful result.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.