ndarray_linalg::tridiagonal

Trait SolveTridiagonalInplace

pub trait SolveTridiagonalInplace<A: Scalar, D: Dimension> {
    // Required methods
    fn solve_tridiagonal_inplace<'a, S: DataMut<Elem = A>>(
        &self,
        b: &'a mut ArrayBase<S, D>,
    ) -> Result<&'a mut ArrayBase<S, D>>;
    fn solve_t_tridiagonal_inplace<'a, S: DataMut<Elem = A>>(
        &self,
        b: &'a mut ArrayBase<S, D>,
    ) -> Result<&'a mut ArrayBase<S, D>>;
    fn solve_h_tridiagonal_inplace<'a, S: DataMut<Elem = A>>(
        &self,
        b: &'a mut ArrayBase<S, D>,
    ) -> Result<&'a mut ArrayBase<S, D>>;
}

Required Methods

fn solve_tridiagonal_inplace<'a, S: DataMut<Elem = A>>( &self, b: &'a mut ArrayBase<S, D>, ) -> Result<&'a mut ArrayBase<S, D>>

Solves a system of linear equations A * x = b tridiagonal matrix A, where A is self, b is the argument, and x is the successful result. The value of x is also assigned to the argument.

fn solve_t_tridiagonal_inplace<'a, S: DataMut<Elem = A>>( &self, b: &'a mut ArrayBase<S, D>, ) -> Result<&'a mut ArrayBase<S, D>>

Solves a system of linear equations A^T * x = b tridiagonal matrix A, where A is self, b is the argument, and x is the successful result. The value of x is also assigned to the argument.

fn solve_h_tridiagonal_inplace<'a, S: DataMut<Elem = A>>( &self, b: &'a mut ArrayBase<S, D>, ) -> Result<&'a mut ArrayBase<S, D>>

Solves a system of linear equations A^H * x = b tridiagonal matrix A, where A is self, b is the argument, and x is the successful result. The value of x is also assigned to the argument.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.